Completed lab

+++ include/tbitfield.h
@@ -9,10 +9,11 @@
 #define __BITFIELD_H__
 
 #include <iostream>
+#include <string>
 
 using namespace std;
 
-typedef unsigned int TELEM;
+typedef uint8_t TELEM;
 
 class TBitField
 {
